requirements/local.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
